% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hd_data.R
\name{generate_hd_data}
\alias{generate_hd_data}
\title{Generate High-Dimensional Data}
\usage{
generate_hd_data(I = 100)
}
\arguments{
\item{I}{number of subjects}
}
\value{
A list with 7 elements, which are:
\describe{
\item{Y}{p-by-J matrix of values}
\item{I}{Total number of subjects}
\item{J}{Total number of observations}
\item{time}{Time of the image collection}
\item{visit}{Vector of number of visits per subjects}
\item{phix0}{a matrix of true phi values for x0}
\item{phix1}{a matrix of truephi values for x1}
\item{phiw}{a matrix of true phi values for w}
\item{xi}{covariate values}
\item{zeta}{true zeta}
}
}
\description{
Generate High-Dimensional Data
}
\examples{
generate_hd_data()
generate_hd_data(I = 20)
}
